pyproject.toml
[project]
name = "repl-nix-docembedassistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.13.3",
    "numpy>=1.26.4",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.41.0",
]
